requirements.txt
marimo==0.11.16
av==14.2.0
pillow==10.4.0
numpy==2.1.1
matplotlib==2.10.1
tqdm==4.67.1
ultralitics==8.3.86
